Refactor View button to use editor hook pattern
- Move View button implementation from navigation-link block to editor hook
- Create navigation-link-view-button.js hook following pattern-overrides.js pattern
- Hook targets core/navigation-link and core/navigation-submenu blocks
- Use __unstableBlockToolbarLastItem to position button as last item
- Remove router dependency issues by using editor hook context
- Register hook in editor hooks index
