+++ packages/editor/src/hooks/index.js
@@ -5,3 +5,4 @@ import './custom-sources-backwards-compatibility';
 import './default-autocompleters';
 import './media-upload';
 import './pattern-overrides';
+import './navigation-link-view-button';
